refactor: move spatial io routines
