+++ pyproject.toml
@@ -86,13 +86,7 @@ omit = [
     "setup.py",
     "conf.py",
     "scripts/*",
-    "pyTMD/calc_astrol_longitudes.py",
-    "pyTMD/check_points.py",
     "pyTMD/compute_tide_corrections.py",
-    "pyTMD/convert_crs.py",
-    "pyTMD/convert_ll_xy.py",
-    "pyTMD/load_constituent.py",
-    "pyTMD/load_nodal_corrections.py",
     "test/_coefficients_to_json.py",
     "test/_def_to_json.py",
     "providers/*.py",
